Move BookmarksPlayerTab to BookmarksListView
